Uso nome bottoni di submit invece di campo "mode" nel post

+++ src/edit_docente.php
@@ -3,8 +3,6 @@ include_once '../database/interface.php';
 include_once 'navigation.php';
 include_once 'token.php';
 
-$mode = filter_var($_POST['mode'], FILTER_SANITIZE_STRING);
-
 $id_docente = intval(filter_var($_POST['id'], FILTER_SANITIZE_NUMBER_INT));
 $nome = mysqli_real_escape_string($connection, $_POST['nome']);
 $cognome = mysqli_real_escape_string($connection, $_POST['cognome']);
@@ -18,32 +16,19 @@ if (!is_amministratore_by_username($token['username'])) {
     go_to_login();
 }
 
-
-switch ($mode) {
-case 'add':
+if (isset($_POST['add'])) {
     try {
         add_docente($nome, $cognome, $username, $password);
     } catch (\Throwable $th) {
         // TODO: mostrare errore in modo migliore
         echo "<h1>Impossibile aggiungere docente: username già esistente</h1>";
         print($th);
-        break;
+        exit();
     }
-    go_to_docenti();
-
-    break;
-
-case 'edit':
+} elseif (isset($_POST['edit'])) {
     edit_docente($id_docente, $nome, $cognome, $username, $password);
-    go_to_docenti();
-    break;
-
-case 'delete':
+} elseif (isset($_POST['delete'])) {
     delete_docente($id_docente);
-    go_to_docenti();
-    break;
-
-default:
-    go_to_docenti();
-    break;
 }
+
+go_to_docenti();
