+++ src/edit_studente.php
@@ -2,8 +2,6 @@
 include_once '../database/interface.php';
 include_once 'navigation.php';
 
-$mode = filter_var($_POST['mode'], FILTER_SANITIZE_STRING);
-
 $id_studente = filter_var($_POST['id_studente'], FILTER_SANITIZE_NUMBER_INT);
 $nome = mysqli_real_escape_string($connection, $_POST['nome']);
 $cognome = mysqli_real_escape_string($connection, $_POST['cognome']);
@@ -17,23 +15,12 @@ if (!is_amministratore_by_username($token['username'])) {
 }
 
 
-switch ($mode) {
-case 'add':
+if (isset($_POST['add'])) {
     add_studente($nome, $cognome, $id_classe);
-    go_to_studenti();
-    break;
-
-case 'edit':
+} elseif (isset($_POST['edit'])) {
     edit_studente($id_studente, $nome, $cognome, $id_classe);
-    go_to_studenti();
-    break;
-
-case 'delete':
+} elseif (isset($_POST['delete'])) {
     delete_studente($id_studente);
-    go_to_studenti();
-    break;
-
-default:
-    go_to_studenti();
-    break;
 }
+
+go_to_studenti();
